feat(6313): make language agnostic parent connector to handle shared functionality
